Add files via upload
A URL crawler which crawls through fetches Wikipedia pages based on the specified topic, which is water conservation and scarcity in this case.
The page names can then be taken from the URLs given by the 'urlcrawler.py' file and entered as a list in 'contentwiki.py'. The output will give us a summary, and limited content from each of the pages.
This approach will reduce time to search for topic-related content in order to fill our application's database.